Refactored architecture
