+++ back/MyStore.Core/Application/IStoreApplication.cs
@@ -1,4 +1,4 @@
-﻿using MyStore.Core.Domain.Dto;
+﻿using MyStore.Core.Data.Dto;
 
 namespace MyStore.Core.Application
 {
