+++ back/MyStore.Core/Application/Mapping.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
-using MyStore.Core.Domain.Model;
-using MyStore.Core.Domain.Dto;
+using MyStore.Core.Data.Model;
+using MyStore.Core.Data.Dto;
 
 namespace MyStore.Core.Application
 {
@@ -27,7 +27,7 @@ namespace MyStore.Core.Application
                     .ConstructUsing(shipping =>
                         new ShippingDto(shipping.Id, shipping.Package, shipping.Price));
             });
-            // only during development, validate your mappings; remove it before release
+            //TOREMOVE
             configuration.AssertConfigurationIsValid();
             return configuration.CreateMapper();
         }
