+++ back/MyStore.Core/Data/Context/Configuration/BaseConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace MyStore.Core.Database.Configuration
+namespace MyStore.Core.Data.Context.Configuration
 {
     public abstract class BaseConfiguration
     {
