+++ back/MyStore.Core/Data/Model/BaseModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace MyStore.Core.Domain
+namespace MyStore.Core.Data.Model
 {
     public abstract class BaseModel
     {
