+++ back/MyStore.Core/Data/Model/Product.cs
@@ -1,4 +1,4 @@
-﻿namespace MyStore.Core.Domain.Model
+﻿namespace MyStore.Core.Data.Model
 {
     public class Product : BaseModel
     {
