+++ back/MyStore.Core/MyStore.Core.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Database\Script\" />
+    <Folder Include="Data\Context\Script\" />
   </ItemGroup>
 
 </Project>
