+++ back/MyStore.UnitTests/BaseTests.cs
@@ -1,4 +1,4 @@
-﻿using MyStore.Core.Database;
+﻿using MyStore.Core.Data.Context;
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Microsoft.DependencyInjection.Abstracts;
