+++ back/MyStore.UnitTests/ShippingTests.cs
@@ -1,6 +1,5 @@
-﻿using MyStore.Core.Database;
-using MyStore.Core.Domain;
-using MyStore.Core.Domain.Model;
+﻿using MyStore.Core.Data.Model;
+using MyStore.Core.Data;
 using MyStore.Core.Repository;
 using System.Collections.Generic;
 using System.Linq;
