+++ back/MyStore.Web/Controllers/ProductsController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using MyStore.Core.Application;
-using MyStore.Core.Domain.Dto;
+using MyStore.Core.Data.Dto;
 
 namespace MyStore.Web.Controllers
 {
